Undo move, better create a commands dir
